🪓 Split SoundBufferData file
